[qt6] Fix compilation warnings

+++ src/BaseAlembicObject.cpp
@@ -22,9 +22,10 @@ void BaseAlembicObject::fillUserProperties(const Alembic::Abc::ICompoundProperty
 
 void BaseAlembicObject::setTransform(const Alembic::Abc::M44d& mat)
 {
-    QMatrix4x4 qmat(mat[0][0], mat[1][0], mat[2][0], mat[3][0], mat[0][1], mat[1][1], mat[2][1],
-                    mat[3][1], mat[0][2], mat[1][2], mat[2][2], mat[3][2], mat[0][3], mat[1][3],
-                    mat[2][3], mat[3][3]);
+    QMatrix4x4 qmat((float)mat[0][0], (float)mat[1][0], (float)mat[2][0], (float)mat[3][0],
+                    (float)mat[0][1], (float)mat[1][1], (float)mat[2][1], (float)mat[3][1],
+                    (float)mat[0][2], (float)mat[1][2], (float)mat[2][2], (float)mat[3][2],
+                    (float)mat[0][3], (float)mat[1][3], (float)mat[2][3], (float)mat[3][3]);
     _transform->setMatrix(qmat);
 }
